+++ src/tkc/async.h
@@ -65,6 +65,20 @@ ret_t async_call(async_exec_t exec, async_on_result_t on_result, void* ctx);
  */
 ret_t async_call_init_ex(uint32_t max_threads, uint32_t min_threads);
 
+/**
+ * @method async_call_init_ex2
+ * 全局初始化。
+ *
+ * @annotation ["static"]
+ * @param {uint32_t} max_threads 最大线程数。
+ * @param {uint32_t} min_threads 最小线程数。
+ * @param {uint32_t} stack_size 栈空间大小(字节)。
+ *
+ * @return {ret_t} 返回RET_OK表示成功，否则表示失败。
+ *
+ */
+ret_t async_call_init_ex2(uint32_t max_threads, uint32_t min_threads, uint32_t stack_size);
+
 #define async_call_init() async_call_init_ex(5, 1)
 
 /**