+++ docs/changes.md
@@ -1,5 +1,8 @@
 # 最新动态
 2022/04/23
+  * 完善 socket\_resolve 防止溢出（感谢叶冬提供补丁）
+  * 完善 tk\_socket\_close 支持AWorks（感谢叶冬提供补丁）
+  * 增加 async\_call\_init\_ex2（感谢叶冬提供补丁）
   * 重构time clock中anchor处理。
   * 重构gauge pointer中anchor处理。